Apply code review feedback from Andres Riancho.
