+++ core/controllers/csp/utils.py
@@ -74,10 +74,10 @@ CSP_DIRECTIVE_VALUE_XSS_FILTER = "filter"
 #Valid Mime Types list
 MIME_TYPES = MimeTypes().types_map[1].values()
 
-#Define NamedTuple tuple subclass to represents a CSP vuln.
+#Define NamedTuple tuple subclass to represents a CSP vulnerability.
 #Declare type here in order to expose it with project visibility
-#and permit cPickle processing to see it,
-#See http://stackoverflow.com/questions/4677012/python-cant-pickle-type-x-attribute-lookup-failed
+#and permit cPickle processing to see it. See link below:
+#http://stackoverflow.com/questions/4677012/python-cant-pickle-type-x-attribute-lookup-failed
 CSPVulnerability = namedtuple('CSPVulnerability', ['desc', 'severity'])            
 
 def site_protected_against_xss_by_csp(response, allow_unsafe_inline=False,
